+++ src/agent.ts
@@ -1,7 +1,12 @@
-import './boot';
+import {boot} from './boot';
 import yargs from 'yargs';
 import Application from './lib/Application';
 import {AgentCoordinator} from './agents/AgentCoordinator';
+import {ApplicationUpdateAgent} from './agents/ApplicationUpdateAgent';
 
-const argv = yargs(process.argv.slice(2)).options({ agent: { type: 'string' } }).argv;
-(async () => await Application.get(AgentCoordinator).start(argv.agent) )();
+(async () => {
+  await boot();
+  const argv = yargs(process.argv.slice(2)).options({ agent: { type: 'string' } }).argv;
+  await Application.get(AgentCoordinator).start(argv.agent)
+  Application.get(ApplicationUpdateAgent).start();
+} )();