+++ src/agents/AgentCoordinator.ts
@@ -2,6 +2,7 @@ import {inject, injectable} from 'inversify';
 import {BasicAgent} from './BasicAgent';
 import {UniswapPoolScannerAgent} from './UniswapPoolScannerAgent';
 import {UniswapPriceScannerAgent} from './UniswapPriceScannerAgent';
+import {UniswapVerificationAgent} from './UniswapVerificationAgent';
 import {Logger} from 'winston';
 
 @injectable()
@@ -12,10 +13,12 @@ export class AgentCoordinator {
   constructor(
     @inject(UniswapPoolScannerAgent) UniswapPoolScannerAgent: UniswapPoolScannerAgent,
     @inject(UniswapPriceScannerAgent) UniswapPriceScannerAgent: UniswapPriceScannerAgent,
+    @inject(UniswapVerificationAgent) UniswapVerificationAgent: UniswapVerificationAgent,
   ) {
     this.agents = {
       UniswapPoolScannerAgent,
       UniswapPriceScannerAgent,
+      UniswapVerificationAgent,
     };
   }