+++ src/server.ts
@@ -1,5 +1,10 @@
-import './boot';
+import {boot} from './boot';
 import Application from './lib/Application';
 import Server from './lib/Server';
+import {ApplicationUpdateAgent} from './agents/ApplicationUpdateAgent';
 
-Application.get(Server).start();
+(async () => {
+  await boot()
+  Application.get(ApplicationUpdateAgent).start();
+  Application.get(Server).start();
+})();