+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 ### Added
+- Uniswap OTA updates
+- Uniswap MongoDB price management 
 - Swagger documentation on /docs.
 - Added RPC Selection before minting
 
@@ -29,6 +31,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [6.0.2] - 2022-01-21
 ### Removed
 - uniswapV3 integration
+>>>>>>> develop
 
 ## [6.0.1] - 2021-12-09
 ### Changed